# src/routes/user/user.controller.ts
import {
  Body,
  ClassSerializerInterceptor,
  Controller,
  Delete,
  Get,
  HttpCode,
  Param,
  Post,
  Put,
  UseInterceptors,
  UsePipes,
} from '@nestjs/common';
import ZodValidationPipe from '../../pipes/zod-validation.pipe';
import { UsersService } from './user.service';
import {
  CreateUserDto,
  createUserSchema,
  UpdatePasswordDto,
  updatePasswordSchema,
  User,
  userIdSchema,
} from '../../dto/user';
import { InstanceNotFoundException } from '../../exceptions/instance-not-found.exception';
import { WrongCurrentPasswordException } from '../../exceptions/wrong-current-password.exception';
import { BadRequestParamsException } from '../../exceptions/bad-request-params.exception';

@Controller('user')
export class UsersController {
  constructor(private readonly usersService: UsersService) {}

  @UseInterceptors(ClassSerializerInterceptor)
  @Get()
  @HttpCode(200)
  async getAllUsers(): Promise<User[]> {
    return this.usersService.getAllUsers();
  }

  @UseInterceptors(ClassSerializerInterceptor)
  @Get(':id')
  @HttpCode(200)
  async getUser(
    @Param('id', new ZodValidationPipe(userIdSchema)) id: string,
  ): Promise<User> {
    const user: User | null = await this.usersService.getUser(id);
    if (!user) throw new InstanceNotFoundException(`user with id = ${id}`);
    return user;
  }

  @UseInterceptors(ClassSerializerInterceptor)
  @Post()
  @HttpCode(201)
  @UsePipes(new ZodValidationPipe(createUserSchema))
  async addUser(@Body() createUserDto: CreateUserDto): Promise<User> {
    /*if (await this.usersService.userWithLoginExists(createUserDto.login))
      throw new BadRequestParamsException(
        'User with this login already exists',
      );*/
    return this.usersService.addUser(createUserDto);
  }

  @UseInterceptors(ClassSerializerInterceptor)
  @Put(':id')
  @HttpCode(200)
  async changeUserPassword(
    @Param('id', new ZodValidationPipe(userIdSchema)) id: string,
    @Body(new ZodValidationPipe(updatePasswordSchema))
    updatePasswordDto: UpdatePasswordDto,
  ): Promise<User> {
    const existingUserRecord: User | null = await this.usersService.getUser(id);
    if (!existingUserRecord) throw new InstanceNotFoundException(`user with id = ${id}`);
    if (existingUserRecord.password !== updatePasswordDto.oldPassword)
      throw new WrongCurrentPasswordException();
    const user: User | null = await this.usersService.updateUserPassword(
      id,
      updatePasswordDto.newPassword,
    );
    if (!user) throw new InstanceNotFoundException(`user with id = ${id}`);
    return user;
  }

  @UseInterceptors(ClassSerializerInterceptor)
  @Delete(':id')
  @HttpCode(204)
  async deleteUser(@Param('id', new ZodValidationPipe(userIdSchema)) id: string) {
    if (!await this.usersService.deleteUser(id))
      throw new InstanceNotFoundException(`user with id = ${id}`);
  }
}